Day 6: Screensaver
Summary:
I like things that are functional and useful, so I wanted to show some useful information in the screensaved (date, battery level, weather).

I started looking online for inspiration, and found many of the old Win screensavers (e.g. the pipes!). At this point my partner hijacked the research and pointed me to this mac screensaver she loved.

Implemented something similar to that, but not quite.. And of course, added some interactivity!