Fix day 6: Day of the week circle
Summary:

Found a bug: day of the week start from 0, so the animation was being triggered every second.

The solution for this problem so far was to add 1 (like for hours, seconds, etc).
I could change things up a bit for a better fix, e.g. not using `metric > 0`, but ain't nobody got time fo' dat (now)